Changes kept or discarded. Spinboxes for patterns
The code keeps track (crudely) of when an LED, flash, or pattern
as been modified. The user is required to explicitly keep or
discard any changes before beginning to edit a different LED, flash,
or pattern.

Spinboxes are used to select the flashes in a pattern, rather than
two sets of radiobuttons.