Improved detection of changed configuration
Added a __eq__ method for all of the firefly data classes. Whenever
the user clicks on a widget we check to see if the class instance
actually changed from the stored configuration value. The user is
only forced to save/discard changes if the instance is different
from the stored version.