Cleaned comments, added Execute frame
Some comments, docstrings, and indenting were cleaned up after
running pylint. The Execute frame was added to execute an LED,
a flash, a pattern, or a pattern set. The Startpage functions
were modified to explicitly close the serial connection. Startpage
was also modified to handle the "Cancel" buttons of the file
selection dialog boxes.